requirements.txt
aioredis==1.3.0
asgiref==3.2.2
asn1crypto==1.1.0
async-timeout==3.0.1
attrs==19.2.0
autobahn==19.10.1
Automat==0.7.0
certifi==2019.9.11
cffi==1.12.3
channel==0.1.0
channels==2.3.0
channels-redis==2.4.0
constantly==15.1.0
cryptography==2.7
daphne==2.3.0
Django==2.2.6
django-cors-headers==3.1.1
djangorestframework==3.10.3
hiredis==1.0.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
msgpack==0.6.2
msgpack-python==0.5.6
pycparser==2.19
PyHamcrest==1.9.0
pytz==2019.3
redis==2.10.6
six==1.12.0
sqlparse==0.3.0
Twisted==19.7.0
txaio==18.8.1
zope.interface==4.6.0
